requirements.txt
streamlit>=1.28.0
langgraph>=0.0.40
langchain-openai>=0.0.5
langchain-community>=0.0.20
langchain-experimental>=0.0.50
langchain-core>=0.1.0
playwright>=1.40.0
python-dotenv>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.8.0
requests>=2.31.0 